requirements.txt
openai
pinecone-client
requests
flask
pytest
jupyter
fastapi 
uvicorn
aiosqlite
python-dotenv